requirements.txt
alembic==1.6.2
asn1crypto==1.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
Flask==1.1.2
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
httpie==3.1.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
multidict==6.0.2
pycparser==2.20
Pygments==2.11.2
PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.14
urllib3==1.26.9
Werkzeug==1.0.1
